requirements.txt
PyExifTool==0.1.1
PyYaml==5.3.1
